Define non-op spiceai change commiter, move stream_changes from flight to spiceai
