+++ crates/data_components/src/cdc.rs
@@ -64,16 +64,13 @@ pub trait CommitChange {
 }
 
 pub struct ChangeEnvelope {
-    change_committer: Option<Box<dyn CommitChange + Send>>,
+    change_committer: Box<dyn CommitChange + Send>,
     pub change_batch: ChangeBatch,
 }
 
 impl ChangeEnvelope {
     #[must_use]
-    pub fn new(
-        change_committer: Option<Box<dyn CommitChange + Send>>,
-        change_batch: ChangeBatch,
-    ) -> Self {
+    pub fn new(change_committer: Box<dyn CommitChange + Send>, change_batch: ChangeBatch) -> Self {
         Self {
             change_committer,
             change_batch,
@@ -81,10 +78,7 @@ impl ChangeEnvelope {
     }
 
     pub fn commit(self) -> Result<(), CommitError> {
-        if let Some(change_committer) = self.change_committer {
-            return change_committer.commit();
-        };
-        Ok(())
+        self.change_committer.commit()
     }
 }
