+++ crates/data_components/src/debezium_kafka.rs
@@ -95,7 +95,7 @@ impl DebeziumKafka {
 
                 let val = msg.value();
                 changes::to_change_batch(&schema, &pk, val)
-                    .map(|rb| ChangeEnvelope::new(Some(Box::new(msg)), rb))
+                    .map(|rb| ChangeEnvelope::new(Box::new(msg), rb))
                     .map_err(|e| cdc::StreamError::SerdeJsonError(e.to_string()))
             });
